+++ src/spy4cast/spy4cast/preprocess_unstructured.py
@@ -87,8 +87,8 @@ class PreprocessUnstructured(_Procedure):
         ns, nt = data.shape
 
         if order is not None and period is not None:
-            b, a = cast(Tuple[npt.NDArray[np.float_], npt.NDArray[np.float_]], signal.butter(order, 2 / period, btype=freq, analog=False, output='ba', fs=None))
-            data = np.apply_along_axis(lambda ts: cast(Tuple[npt.NDArray[np.float_]], 
+            b, a = cast(Tuple[npt.NDArray[np.float32], npt.NDArray[np.float32]], signal.butter(order, 2 / period, btype=freq, analog=False, output='ba', fs=None))
+            data = np.apply_along_axis(lambda ts: cast(Tuple[npt.NDArray[np.float32]], 
                                                        signal.filtfilt(b, a, ts)), 1, data)
         elif order is not None or period is not None:
             if order is None:
@@ -219,7 +219,7 @@ class PreprocessUnstructured(_Procedure):
         return self._land_data
 
     @property
-    def data(self) -> npt.NDArray[np.float_]:
+    def data(self) -> npt.NDArray[np.float32]:
         """Raw data in the object with `nan` as in the original dataset. It has dimensions of
         space x time."""
         return self._land_data.values
